Refactoring of things adjacent to entity factory (#759)
With the decompiling of the entity factory framework being started, I
went through and replaced some placeholders that have accumulated over
time. Originally I was going to try to decompile the whole sequence of
events that happens when you cast Hellfire, but I think the core of that
is in EntityAlucard, since we set a player step.

Overall, this should have no functional changes and should all just be
renaming things to improve based on new knowledge.