Use FakePrim on EntityEnemyBlood (#793)
After decompiling the Vibhuti in my last PR, I looked at the pattern I
had noticed, where users of FakePrim would call `func_800EDB58`. I then
looked to that function's callers:
https://raw.githubusercontent.com/Xeeynamo/sotn-decomp/gh-duplicates/function_calls/dra.func_800EDB58.svg

Clearly, EntityEnemyBlood uses this function. Sure enough, when I
examined that function in its current state, I found that it has a bunch
of ugly casts which directly match FakePrim. Therefore, this PR changes
it to use FakePrim so that it does not need all those casts anymore.
Additionally, we find in this function that the velocity is being
changed by another member in the struct, which tells us that this member
is acceleration. Therefore, FakePrim is being understood more and more.