Refactor some functions with gravity jump (#815)
- Rename func_800FF064 to HandleGravityBootsMP (call graph proves that
this is the only use:
https://raw.githubusercontent.com/Xeeynamo/sotn-decomp/gh-duplicates/function_calls/dra.func_800FF064.svg)
- Rewrite that function to use the CallMode system we developed earlier,
and also rewrite it to not need the extra variable. Function is now much
shorter.
- Rename func_8010E9A4 to DoGravityJump. This function sets Alucard's
velocity to move up rapidly, spawns the gravity boot laser beam entity,
and sets the player state machine's step.
- Fill in enums for SetPlayerStep in a few related places.

Overall I think this makes the logic more readable. Next going to look
into g_Player.unk44 because that seems to be the last missing piece
here.