Decompile RIC func_80156DE4 (#817)
This one is a bit frustrating. The reason for that is that this is a
duplicate of DRA `func_80110394`. Luckily, when I started decompiling
this one, I recognized the assembly as being similar, since I decompiled
that function in DRA a long time ago. I think the reason the function
duplicate finder misses it is that it uses g_api.CheckCollision, while
the DRA version was able to directly call CheckCollision without going
through g_api.

Also, this one uses `i<3` in its for-loop, while the DRA one used i<4.

Overall, seems that being aware of `g_api` mismatches will be important
to keep in mind as we go forward. I think I'm going to try to decompile
a lot of RIC in the near future, so that it can catch up to DRA. And
maybe if we're lucky, that will help to solve some of the functions that
are currently tricky in DRA. We'll see.