Decompile RIC func_8017091C (#820)
This appears to be an entity updating function unique to RIC. I do not
know what it is for yet.

Main reason for doing it is that it has rodata and the file it is in
currently requires a rodata padding to be present. Having this one
working and parsing the rodata into it means we can move the padding,
and once we finish everything in this file, we will be able to remove
the padding entirely.