Sound sim: Add save_mgr, 4A538, collider, demo, menu (#839)
Add a bunch of source files into the sound test app.

I implemented `firstfile`, `nextfile` and `erase`. To make them work
properly `_adjust_path` needs to be implemented. I left an `INFOF` as a
reminder.

`func_800EDD9C` is matching but the implementation was completely wrong.
I re-decompiled it appropiately and removed some fake symbols.

`SetPolyGT4` and friends are ticking bombs since on different platforms
than PS1 it is not possible to freely play with pointers. The struct
`P_TAG` would need to be rewritten.

`DEMO_KEY_PTR` is stubbed for the same reason above.