Add File specific funcs (#898)
I am not sure if we want to use this pattern. Here I am proposing to
wrap functions to open files and read the entirety of a file.

The objective is to avoid performing sanity checks for `fopen`, `fread`
and `malloc`, other than ensuring allocated resources are correctly
de-allocated on both happy and unhappy paths.

How it works is the first parameter is always the callback, while
starting from the second are all the parameters necessary to instantiate
and allocate the resources to be passed to the callback.