Decompile RIC EntityShrinkingPowerUpRing (#973)
Happy to rename this one but this is the best I came up with.

This is the entity function for the faint rings that appear around
Richter in two cases. 1: When he gets saved by Maria after dying in the
prologue, and 2: When he powers up for the flame-whip ability when you
press triangle without a subweapon.

Here is a screenshot of the former case: 

![image](https://github.com/Xeeynamo/sotn-decomp/assets/15314202/3719d784-c30c-4c4f-8d51-6662fbb93d5f)

This is actually 4 instances of the ring, nested within each other. I'll
be honest, I never even noticed this existed. This is a huge amount of
code for something so minor.