Decompile RIC func_80167A70 (#989)
As noted in the comment, I don't know what this actually is and how it
corresponds to DRA, but hopefully the comment will help track it down in
the future.

In the entity extension, this function just has a timer at offset 7C.
The gravity boot beam entity had the same. I decided to create an
extension called TimerOnly to handle these multiple entities that just
have a timer there. I think finding patterns like these in the entities
will be nice to try to slim down entity.h.